Allow randomizing s3 bucket name & update python run time (#6)
Randomizing s3 bucket name is essential because s3 bucket names must be globally unique across the entire aws. This means that a bucket name such as metaflow-s3-dev is unlikely to qualify, hence we add an optional switch to the module that if set will append a random 8-character string into the bucket name to make it qualify.

The python runtime upgrade is necessary as 3.7 is deprecated by lambda and creating a new lambda fails.