Add optional ALB support to metadata service
This PR adds a `setup_alb` variable to metadata service module which if set also sets up an application load balancer in front of the metadata service. The network load balancer setup previously would also remain active so that 2 load balancers would be serving the same ECS service. Eventually the plan is to completely deprecate and remove the NLB but to do it with zero-downtime we will first setup the ALB, then redirect all the traffic slowly to the ALB instead of NLB and then remove the NLB altogether.

Test plan
---
Test via cdktf plan